move FilterRow to own file
